requirements.txt
python-fasthtml
python-dotenv
sqlalchemy
psycopg2-binary
bcrypt
PyJWT
cryptography
langchain-openai
langgraph
httpx
pdfplumber
plotly
